chore_: reorganize thirdparty dir into categories
