+++ services/wallet/decoder.go
@@ -2,8 +2,8 @@ package wallet
 
 import (
 	"github.com/status-im/status-go/services/wallet/thirdparty"
-	"github.com/status-im/status-go/services/wallet/thirdparty/fourbyte"
-	"github.com/status-im/status-go/services/wallet/thirdparty/fourbytegithub"
+	"github.com/status-im/status-go/services/wallet/thirdparty/decoder/fourbyte"
+	"github.com/status-im/status-go/services/wallet/thirdparty/decoder/fourbytegithub"
 )
 
 type Decoder struct {
