+++ services/wallet/service.go
@@ -33,11 +33,11 @@ import (
 	"github.com/status-im/status-go/services/wallet/router"
 	"github.com/status-im/status-go/services/wallet/router/pathprocessor"
 	"github.com/status-im/status-go/services/wallet/thirdparty"
-	"github.com/status-im/status-go/services/wallet/thirdparty/alchemy"
-	"github.com/status-im/status-go/services/wallet/thirdparty/coingecko"
-	"github.com/status-im/status-go/services/wallet/thirdparty/cryptocompare"
-	"github.com/status-im/status-go/services/wallet/thirdparty/opensea"
-	"github.com/status-im/status-go/services/wallet/thirdparty/rarible"
+	"github.com/status-im/status-go/services/wallet/thirdparty/collectibles/alchemy"
+	"github.com/status-im/status-go/services/wallet/thirdparty/collectibles/opensea"
+	"github.com/status-im/status-go/services/wallet/thirdparty/collectibles/rarible"
+	"github.com/status-im/status-go/services/wallet/thirdparty/market/coingecko"
+	"github.com/status-im/status-go/services/wallet/thirdparty/market/cryptocompare"
 	"github.com/status-im/status-go/services/wallet/token"
 	"github.com/status-im/status-go/services/wallet/transfer"
 	"github.com/status-im/status-go/services/wallet/walletevent"
